remove files that are not used, split out and rename node ws components
