+++ fork-visualizer/src/app/components/NodeWsController.jsx
@@ -4,7 +4,7 @@ import { ApiPromise, WsProvider } from "@polkadot/api";
 import { useState } from "react"
 
 const defaultFunc = () => console.log("oops this is the default func")
-export function NodeState ({ wsAddress, updateStuff, removeNode }) {
+export default function NodeWsController ({ wsAddress, updateStuff, removeNode }) {
 
     const [ isSubscribed, setIsSubscribed ] = useState(false)
     const [ loading, setLoading ] = useState(false)
@@ -74,45 +74,5 @@ export function NodeState ({ wsAddress, updateStuff, removeNode }) {
             </th>
             <th className="px-4">{ latestHash && <span>{ latestHash }</span> }</th>
         </tr>
-    )
-}
-
-
-export default function NodeTracker ({ addNode , children}) {
-
-    const [ nodeAddr, setNodeAddr ] = useState("")
-
-    const onAddNode = () => {
-        addNode(nodeAddr)
-        setNodeAddr("")
-    }
-
-    const onKeyDownHandler = e => {
-        if (e.keyCode === 13) {
-            onAddNode()
-            setNodeAddr("")
-        }
-    }
-
-    return (
-        <div className="fixed z-50 flex flex-col top-1 left-1">
-            <div className="flex flex-row gap-2">
-                <input
-                    className="px-2 py-1 my-1 rounded-md"
-                    type="text"
-                    name="nodeAddr"
-                    value={nodeAddr}
-                    placeholder="WS address to subscribe to"
-                    onChange={(e) => {
-                        setNodeAddr(e.target.value)
-                    }}
-                    onKeyDown={onKeyDownHandler}
-                />
-
-                <button className="text-white btn" onClick={onAddNode}>Add</button>
-            </div>
-
-            { children }
-        </div>
     )
 }
